sprite and texture index hash tables
indices of sprites and textures are now stored in a hash table. this
makes loading them on initialization a lot more intuitive; also
practically required for adding connected textures soon :)

+++ src/entity.c
@@ -6,7 +6,7 @@
 #include "world.h"
 #include "list.h"
 
-entity_t *entity_create(int sprite, v3d pos, v3d size) {
+entity_t *entity_create(size_t sprite, v3d pos, v3d size) {
 	entity_t *entity = (entity_t *)malloc(sizeof(entity_t));
 
 	entity->sprite = sprite;
